Adding Promise polyfill to Session class for MSIE
Internet Explorer and Opera Mini are the only browsers than don't support promises and adding this small polyfill addresses that problem (and fixes #9).

There are other polyfills that are better in scope, but at this point only Internet Explorer needs it this is a simple and compatible solution.

Note: Some Promise related warnings (but not errors) may be still generated in the console in MSIE which using a different polyfill might resolve, but this does not impact the user experience.