Email server now configurable via .ENV
This fixes #27 which is that the live demo site was not longer sending email (it was being blocked, possiibly because it was sending too many as the demo and/or now.sh was getting popular and hit rate limits, or because it now.sh change their policy and started blocking email).

You can now easily configure the project to use a mail server to send email reliably. This will make it easier for people to try the demo and see that it works.