/auth/profile is now /account
Refactored so that the redirection on signin is handled only in once place, in ./pages/auth/signin (previously it was handled sepreately client and server side in multiple files).

This reduces complexity and makes changing the page users are re-directed to on sign in easier as it only needs to be changed in once place now.

This also makes making process on #52 easier if that’s something we want to pick up on in future.